temp fix for stringify not converting MetaString
When metadata is stored in the metadata section of a defaults file,
some of the strings are stored as MetaString objects, instead of
MetaInline (which is the case when metadata comes from the document
header).

Unfortunately, pandoc's stringify currently does not work on MetaString
objects, so this provides a temporary workaround. (This issue has been
reported upstream at https://github.com/jgm/pandoc/issues/7149.)